Move plugin loading to where it belongs
The plugin loading code, instead of being in the DiscordReady event, has
accidentally been placed in the message event. This caused plugins not
to load until the first message arrives.

This has been fixed by moving the code into the proper place at the end
of the `ready` event handler.