Prevent crash when trying to list an invalid plugin
There was no error catching before when listing plugins, which led to a
crash of the bot every time someone tries to list all plugins in the
plugin directory if it contains a plugin that is in some shape or form
not loadable.

A try/catch prevents this from happening and tells the user that the
plugin has errored out.