Add Version Check, make version depend on `package.json`
A version check now checks if the bot is up to date every time it starts
up and prints a message to the console if this is not the case, asking
the user to update.
It also sets a value in the API for plugins to access so they can verify
if the bot is running the latest version as well. If an update is
available, there will also be a `LATEST_VERSION` value in the API for
plugins to compare.

The second part of this commit makes all the versions in `app.js`
dependant on the version in `package.json`, which makes for a more
manageable update process.