Add main application
With all the modules in place, there needs to be a main application that
handles the gluing those single pieces together.

The main application does the config loading, initiation of the modules,
connects to the discord network and provides the Api for plugins to use.
It also does the plugin loading and hooks a few simple commands for the
bot's stdin.